[POC] [modern] checkboxHack: Persist sidebar collapsed state client-side also when logged out
Saves the state in LocalStorage, if available.

The loading (init) script is run synchronously before the sidebar is rendered
to avoid the sidebar flashing.
The saving script is loaded later, asynchronously.
The user might toggle the sidebar before the saving logic is loaded,
then the state will be saved as soon as the JS is loaded.

On slow connections this delay might be enough for the user to reload the page
in which case the state would not be saved. This use-case however
is not a practical usage, therefore it's an accepted trade-off.

Based on: Ib1ce934f3646cd8feebf0d3b15c38b5b969ec957

Bug: T257169
Depends-On: I8f336c2f551908cd28247e8f3e75c0e06aa9e23e
Change-Id: I95c01e5797fdeaef12df179e260551407c8302a6